# Faster info via meta class