--- a/astropy/table/tests/test_info.py
+++ b/astropy/table/tests/test_info.py
@@ -248,14 +248,6 @@ def test_no_deprecation_warning():
         assert len(warns) == 0
 
 
-# No problem any more?
-# def test_lost_parent_error():
-#     c = table.Column([1, 2, 3], name='a')
-#     with pytest.raises(AttributeError) as err:
-#         c[:].info.name
-#     assert 'failed access "info" attribute' in str(err.value)
-
-
 def test_info_serialize_method():
     """
     Unit test of context manager to set info.serialize_method.  Normally just