@taldcroft - I decided to try the metaclass approach to `DataInfo`, where for every attribute one automatically generates a descriptor that either gets/sets in `_attrs` or from the parent. It is a bit of magic in the meta-class, but not really that much more than there was before in `__getattr__` and `__setattr__`. Note that I had to move to `__slots__` in order to continue to disallow setting arbitrary attributes; not completely clear whether that is worth it. I have some other comments in-line.

Comparing this PR to current master:
```
from astropy.table import MaskedColumn
m = MaskedColumn([1, 2])
%timeit m.info
# 1000000 loops, best of 5: 290 ns per loop
# on current master: 1.24 µs
%timeit m.info.name
# 1000000 loops, best of 5: 659 ns per loop
# on current master: 2.12 µs

from astropy import units as u
q = [1, 2] * u.m
%timeit q.info
# 1000000 loops, best of 5: 319 ns per loop
# on current master: 1.82 µs
%timeit q.info.name
# 1000000 loops, best of 5: 463 ns per loop 
# on current master: 2.46 µs
```

The last item is the most relevant for possibly moving everything to `info` - `463 ns` is still about 6 times slower than just `%timeit m.name`. But it is not obvious to me that with the general speed-up here, it would still be a limiting factor.