@taldcroft 2019-07-15
@mhvk - this looks great from the description, but clearly needs a decent chunk of time to dig into and evaluate.  So it's on my radar, but probably not in the next few days...