@taldcroft 2019-07-22
@mvhk - have you had a look at https://www.python.org/dev/peps/pep-0487/ ?  This is available as of Python 3.6.  Not clear if the `__slots__` stuff is possible, but it would be nice to avoid metaclasses if possible.