@mhvk 2019-07-22
(Taking this out of the inline comments as I think it goes to the heart of the matter.)

@taldcroft - The approach with just defining descriptors on the class is indeed a lot clearer! 

The main reason I went for the meta-class approach was to keep the API similar, with its `attr_names` and `attrs_from_parent` - which is why there are relatively few changes to `Info` subclasses.

The question then is whether we are happy to change it, which in turn depends on whether anybody outside of astropy relies on it. In principle, I guess the metaclass could help here, by doing the auto-generation while emitting a deprecation warning. But of course easier if we can do without it...