@mhvk 2019-07-22
p.s. Have been meaning to look at the effects of PEP 487... Definitely useful for simple registration like Time fornats.