@taldcroft 2019-07-23
> .. was to keep the API similar, with its attr_names and attrs_from_parent

Ah yes, I didn't immediately think of the API.  That's possibly thorny... So what about some radical thoughts just for the sake of conversation:

- What if we left the existing `DataInfo` hierarchy in place for external usage (but deprecate) and then made a brand new infrastructure that keeps most of the public API but is not constrained to implement in the same way.  For the sake of discussion all it `XDataInfo` or `XMixinInfo`.
- `attr_names` and `attrs_from_parents` can be done with `__set_name__` in the descriptors.
- Basically saying, `Info` is a honking good idea, let's make it better and faster with a reboot.  In practice it probably won't be radically different, but just a chance to take stock of what is there and make it nicer.
- I *think* that in the rest of astropy there are not too many places we would need to change to reflect this additional `XDataInfo` class hierarchy (but I haven't tried looking).
- Our Astropy classes would have e.g. `class QuantityInfoBase(XParentDtypeInfo)`, `class TimeInfo(XMixinInfo)`, etc.