@taldcroft 2019-07-23
Having given a decent first pass through this code and seen the timing, it looks great!  So it might be a reasonable strategy to get this in shape to merge and separately think about a bigger redesign to make the code more elegant / readable and maybe even faster.