@mhvk 2019-07-23
OK, I've implemented the suggested changes. I like the idea of merging this, and worrying about a simpler design separately. The one outstanding question, I think, is whether it is worth the hassle with `__slots__` to prevent people from sticking arbitrary attributes on `info`.