@taldcroft 2019-07-27
Here are some timing results on my laptop with your branch and master.  I get qualitatively similar results.  This also explores the speedup of a different strategy (attrs are stored as object attributes, not dict elements).

```python
from weakref import ref as weakref
from astropy.utils.data_info import DataInfo
from astropy.table import MaskedColumn
```


```python
%astro
```

    astropy=4.0.dev25224



```python
DataInfo._attr_defaults  # Shows this is the faster-info-via-meta-class branch
```




    {'dtype': dtype('O')}




```python
class A:
    info = DataInfo()
```


```python
a = A()
```


```python
a.info.description = 'asdf'
```


```python
timeit a.info
```
    463 ns ± 3.27 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
MASTER:  2.5 µs ± 43.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)


```python
timeit a.info.description
```

    845 ns ± 13.7 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
 MASTER: 3.39 µs ± 64.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)


```python
a.__dict__
```




    {'info': dtype = object
     shape = --
     description = asdf
     class = A
     n_bad = 0}




```python
class SimpleDataInfo:
    """Simplest possible data info container"""
    __slots__ = ['name', 'description', 'dtype', 'meta', 'unit', '_parent_ref']
```


```python
class B:
    """Use SimpleDataInfo via a class property"""
    @property
    def info(self):
        try:
            info = self._info
        except AttributeError:
            info = self._info = SimpleDataInfo()
        finally:
            # info._parent_ref = self # weakref(self)
            return info            
```


```python
b = B()
```


```python
b.info.name = 'asdf'
```


```python
timeit b.info.name
```

    207 ns ± 6.02 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)



```python
class FDataInfo:
    """Descriptor that roughly matches the property above"""
    def __get__(self, instance, owner_cls):
        try:
            info = instance._info
        except AttributeError:
            if instance is None:
                # This is an unbound descriptor on the class
                self._parent_cls = owner_cls
                return self
            else:
                info = instance._info = SimpleDataInfo()
        finally:
            # info._parent_ref = weakref.ref(instance)
            return info
```


```python
class C:
    info = FDataInfo()
```


```python
c = C()
```


```python
c.info.description = 'hello'
```


```python
timeit c.info.description
```

    231 ns ± 4.75 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)



```python
m = MaskedColumn([1, 2])
m.info.description = 'asdf'
```


```python
timeit m.description
```

    50 ns ± 0.801 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)



```python
timeit m.info.description
```

    1.01 µs ± 10.2 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)



```python
junk = SimpleDataInfo()
m.junk = junk
```


```python
timeit m.junk
```

    58.1 ns ± 0.271 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)



```python
m.junk.name = 'name'
```


```python
timeit m.junk.name
```

    86.1 ns ± 0.644 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)