@taldcroft 2019-08-29
:+1: on explicitly preventing people from setting arbitrary attributes on `info`.  That matches the current release behavior and I think it prevents mistakes far more often than being a limitation.  Along those lines, I have often done things like `q.name = name` on a Quantity when I meant `q.info.name`.

The current implementation as slots is thus fine to me.  If we even move to a different implementation that does the same thing then that will also be fine.