@mhvk 2019-09-14
OK, I reinstated the test and fixed the code so that the appropriate message is given again. The difference was that now there always is a `weakref` - it can just be dead, in which case `_parent_ref()` returns `None`.